rename RTDB components
