+++ packages/flutterfire_ui/example/lib/stories.dart
@@ -5,6 +5,8 @@ import 'package:flutterfire_ui_example/stories/screens/sms_code_input_screen.dar
 import 'package:flutterfire_ui_example/stories/views/forgot_password_view.dart';
 import 'package:flutterfire_ui_example/stories/widgets/apple_sign_in_button.dart';
 import 'package:flutterfire_ui_example/stories/widgets/facebook_sign_in_button.dart';
+import 'package:flutterfire_ui_example/stories/widgets/firebase_database_list_view.dart';
+import 'package:flutterfire_ui_example/stories/widgets/firebase_table.dart';
 import 'package:flutterfire_ui_example/stories/widgets/firestore_list_view.dart';
 import 'package:flutterfire_ui_example/stories/widgets/firestore_table.dart';
 import 'package:flutterfire_ui_example/stories/widgets/google_sign_in_button.dart';
@@ -69,6 +71,8 @@ class StoriesApp extends StatelessWidget {
           SignOutButtonStory(),
           FirestoreListViewStory(),
           FirestoreTableStory(),
+          FirebaseDatabaseListViewStory(),
+          FirebaseDatabaseTableStory(),
         ],
       ),
     );
