+++ packages/flutterfire_ui/lib/database.dart
@@ -1,10 +1,11 @@
-export 'src/firebase_query_builder.dart'
+export 'src/firebase_database/query_builder.dart'
     show
-        FirebaseQueryBuilder,
-        FirebaseListView,
-        QueryBuilderSnapshot,
+        FirebaseDatabaseQueryBuilder,
+        FirebaseDatabaseListView,
+        FirebaseQueryBuilderSnapshot,
         FirebaseErrorBuilder,
         FirebaseItemBuilder,
         FirebaseLoadingBuilder;
 
-export 'src/firebase_table_builder.dart' show FirebaseDataTable;
+export 'src/firebase_database/table_builder.dart'
+    show FirebaseDatabaseDataTable;
