+++ packages/flutterfire_ui/lib/firestore.dart
@@ -1,11 +1,11 @@
-export 'src/query_builder.dart'
+export 'src/firestore/query_builder.dart'
     show
         FirestoreQueryBuilder,
-        QueryBuilderSnapshot,
+        FirestoreQueryBuilderSnapshot,
         FirestoreListView,
         FirestoreErrorBuilder,
         FirestoreItemBuilder,
         FirestoreLoadingBuilder,
         FirestoreQueryBuilderSnapshotBuilder;
 
-export 'src/firestore_table_builder.dart' show FirestoreDataTable;
+export 'src/firestore/table_builder.dart' show FirestoreDataTable;
